requirements.txt
requests
python-daemon-3k